Make backend send permitted faction leader list
Make the backend send a list of permitted senators to the frontend when generating select faction leader potential actions.
